+++ QuadricsBasedMeshSimplifier/vertex.h
@@ -24,7 +24,7 @@ public:
 	typedef std::vector<Face*> FaceListType;
 
 public:
-	Vertex(const VR::Vertex& parVertex, size_t parId);
+	Vertex(const VR::Vertex& parVertex, uint parId);
 	~Vertex();
 
 public:
@@ -47,12 +47,12 @@ public:
 	inline const VR::Vec4& Pos() const { return pos_; }
 	inline const FaceListType& IncidentFaces() const { return incidentFaces_; }
 	inline const Quadric& GetQuadric() const { return quadric_; }
-	inline size_t Id() const { return id_; }
+	inline uint Id() const { return id_; }
 	inline const PairListType& Pairs() const { return pairs_; }
 	inline bool DeleteMe() const { return deleteMe_; }
 
 public:
-	inline void SetId(size_t parId) { id_ = parId; }
+	inline void SetId(uint parId) { id_ = parId; }
 	inline void SetDeleteMe() { deleteMe_ = true; }
 	inline void SetPos(const VR::Vec4& parPos) { pos_ = parPos; }
 #ifdef _DEBUG
@@ -68,7 +68,7 @@ private:
 	VR::Vec4 pos_;
 	Quadric quadric_;
 	
-	size_t id_;
+	uint id_;
 	bool deleteMe_;
 };
 // ----------------------------------------------------------------------------
